Bugfix: new packages should be downloaded in the base layer
The use case for it is that the user might want to use
`add-packages-to-pool` where the packages come from extra (non-ubuntu)
repositories.

In my case, I setup the repositories in the base layer and notices they
were not used in a further `add-package-to-pool` action, as `add-package-to-pool`
was being executed in a different layer.

The Ubuntu Server version I am using is 22.04.